+++ CaptchaSharp/Services/DeathByCaptchaService.cs
@@ -211,7 +211,7 @@ namespace CaptchaSharp.Services
 
         #region Reporting the solution
         public async override Task ReportSolution
-            (int id, CaptchaType type, bool correct = false, CancellationToken cancellationToken = default)
+            (long id, CaptchaType type, bool correct = false, CancellationToken cancellationToken = default)
         {
             if (correct)
                 throw new NotSupportedException("This service doesn't allow reporting of good solutions");
